+++ ui/js/component/fileActions/index.js
@@ -18,8 +18,16 @@ import {
 } from "selectors/claims";
 import { doPurchaseUri, doLoadVideo, doStartDownload } from "actions/content";
 import FileActions from "./view";
+import { makeSelectClaimForUri } from "selectors/claims";
+import {
+  doClaimNewSupport,
+  doSetClaimSupportAmount,
+  doSetClaimSupportClaim,
+} from "actions/claims";
+import { selectClaimSupportAmount } from "selectors/claims";
 
 const makeSelect = () => {
+  const selectClaim = makeSelectClaimForUri();
   const selectFileInfoForUri = makeSelectFileInfoForUri();
   const selectIsAvailableForUri = makeSelectIsAvailableForUri();
   const selectDownloadingForUri = makeSelectDownloadingForUri();
@@ -29,6 +37,7 @@ const makeSelect = () => {
   const selectClaimForUri = makeSelectClaimForUri();
 
   const select = (state, props) => ({
+    claim: selectClaim(state, props),
     fileInfo: selectFileInfoForUri(state, props),
     /*availability check is disabled due to poor performance, TBD if it dies forever or requires daemon fix*/
     isAvailable: true, //selectIsAvailableForUri(state, props),
@@ -39,6 +48,7 @@ const makeSelect = () => {
     loading: selectLoadingForUri(state, props),
     claimIsMine: selectClaimForUriIsMine(state, props),
     claimInfo: selectClaimForUri(state, props),
+    amount: selectClaimSupportAmount(state),
   });
 
   return select;
@@ -53,6 +63,10 @@ const perform = dispatch => ({
   startDownload: uri => dispatch(doPurchaseUri(uri, "affirmPurchase")),
   loadVideo: uri => dispatch(doLoadVideo(uri)),
   restartDownload: (uri, outpoint) => dispatch(doStartDownload(uri, outpoint)),
+  claimNewSupport: () => dispatch(doClaimNewSupport()),
+  setAmount: event => dispatch(doSetClaimSupportAmount(event.target.value)),
+  setClaimSupport: (claim_id, name) =>
+    dispatch(doSetClaimSupportClaim(claim_id, name)),
 });
 
 export default connect(makeSelect, perform)(FileActions);